slight code improvements (idk why)

+++ src/minecraft/client/rapid/util/MathUtil.java
@@ -3,9 +3,10 @@ package client.rapid.util;
 public class MathUtil {
 
     public static boolean isInPercentage(double value) {
-        return (int) randomNumber(1, 10) < value;
+        return randomNumber(1, 10) < value;
     }
-	
+
+    // Get a random number between max and min
     public static double randomNumber(double max, double min) {
         return (Math.random() * (max - min)) + min;
     }
